Categorize's Function/requirements.txt
tensorflow
transformers[tf-cpu]
google-cloud-storage
pickle-mixin
gunicorn